implemented logout function
