+++ src/main/java/lt/academy/javau5/pizza/_security/dto_response/MsgResponse.java
@@ -1,10 +1,12 @@
 package lt.academy.javau5.pizza._security.dto_response;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
+@Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class MsgResponse implements AbstractResponse {
