+++ src/main/java/lt/academy/javau5/pizza/_security/dto_response/MultiErrorResponse.java
@@ -1,11 +1,13 @@
 package lt.academy.javau5.pizza._security.dto_response;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import java.util.List;
 
 @Data
+@Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class MultiErrorResponse implements AbstractResponse {
