+++ src/main/java/lt/academy/javau5/pizza/_security/services/HttpResponseService.java
@@ -16,8 +16,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import static org.springframework.http.HttpStatus.*;
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
+import static org.springframework.http.HttpStatus.NOT_ACCEPTABLE;
 
 @Service
 @RequiredArgsConstructor
