+++ src/main/java/lt/academy/javau5/pizza/_security/services/LogoutService.java
@@ -1,41 +1,41 @@
-package lt.academy.javau5.pizza._security.services;
-
-import jakarta.servlet.http.Cookie;
-import lt.academy.javau5.pizza._security.repositories.TokenRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.logout.LogoutHandler;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class LogoutService implements LogoutHandler {
-  private final TokenService tokenService;
-  private final JwtService jwtService;
-
-  @Override
-  public void logout(
-      HttpServletRequest request,
-      HttpServletResponse response,
-      Authentication authentication
-  ) {
-    final String authHeader = request.getHeader("Authorization");
-    final String jwt;
-    if (authHeader == null ||!authHeader.startsWith("Bearer ")) {
-      return;
-    }
-    jwt = authHeader.substring(7);
-    var storedToken = tokenService.findByToken(jwt)
-        .orElse(null);
-    if (storedToken != null) {
-      storedToken.setExpired(true);
-      storedToken.setRevoked(true);
-      tokenService.save(storedToken);
-      jwtService.deleteRefreshTokenCookie(response);
-      SecurityContextHolder.clearContext();
-    }
-  }
-}
+//package lt.academy.javau5.pizza._security.services;
+//
+//import jakarta.servlet.http.Cookie;
+//import lt.academy.javau5.pizza._security.repositories.TokenRepository;
+//import jakarta.servlet.http.HttpServletRequest;
+//import jakarta.servlet.http.HttpServletResponse;
+//import lombok.RequiredArgsConstructor;
+//import org.springframework.security.core.Authentication;
+//import org.springframework.security.core.context.SecurityContextHolder;
+//import org.springframework.security.web.authentication.logout.LogoutHandler;
+//import org.springframework.stereotype.Service;
+//
+//@Service
+//@RequiredArgsConstructor
+//public class LogoutService implements LogoutHandler {
+//  private final TokenService tokenService;
+//  private final JwtService jwtService;
+//
+//  @Override
+//  public void logout(
+//      HttpServletRequest request,
+//      HttpServletResponse response,
+//      Authentication authentication
+//  ) {
+//    final String authHeader = request.getHeader("Authorization");
+//    final String jwt;
+//    if (authHeader == null ||!authHeader.startsWith("Bearer ")) {
+//      return;
+//    }
+//    jwt = authHeader.substring(7);
+//    var storedToken = tokenService.findByToken(jwt)
+//        .orElse(null);
+//    if (storedToken != null) {
+//      storedToken.setExpired(true);
+//      storedToken.setRevoked(true);
+//      tokenService.save(storedToken);
+//      jwtService.deleteRefreshTokenCookie(response);
+//      SecurityContextHolder.clearContext();
+//    }
+//  }
+//}
