+++ src/main/resources/application.properties
@@ -29,6 +29,5 @@ spring.messages.basename=Language/messages
 # Security
 # ===============================
 application.security.jwt.secret-key=404E635266556A586E3272357538782F413F4428472B4B6250645367566B5970
-#application.security.jwt.expiration=86400000
 application.security.jwt.expiration=86400000
 application.security.jwt.refresh-token.expiration=604800000
